# src/main/resources/openapi/causa-api.yaml
openapi: 3.0.3
info:
  title: Causa API
  version: 0.2.0
  description: API for issue tracking and phase workflows.
servers:
  - url: /
paths:
  /issues:
    post:
      tags:
        - Issues
      summary: Create a new issue
      operationId: createIssue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: Issue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
    get:
      tags:
        - Issues
      summary: List issues
      operationId: listIssues
      parameters:
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: assigneeId
          in: query
          required: false
          schema:
            type: string
        - name: memberId
          in: query
          required: false
          schema:
            type: string
        - name: projectId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/IssueStatus'
        - name: phaseKind
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PhaseKind'
      responses:
        '200':
          description: Issue list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IssueListItem'
  /issues/similar:
    get:
      tags:
        - Issues
      summary: Find similar issues
      operationId: findSimilarIssues
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Similar issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IssueListItem'
  /issues/facets:
    get:
      tags:
        - Issues
      summary: Faceted issue filters
      operationId: getIssueFacets
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
        - name: assigneeId
          in: query
          required: false
          schema:
            type: string
        - name: memberId
          in: query
          required: false
          schema:
            type: string
        - name: projectId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/IssueStatus'
        - name: phaseKind
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PhaseKind'
      responses:
        '200':
          description: Faceted options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueFacetResponse'
  /issues/{issueId}:
    get:
      tags:
        - Issues
      summary: Get issue by id
      operationId: getIssue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Issue details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '404':
          description: Issue not found
    patch:
      tags:
        - Issues
      summary: Update issue metadata
      operationId: updateIssue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIssueRequest'
      responses:
        '200':
          description: Issue updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '404':
          description: Issue not found
  /issues/{issueId}/owner:
    patch:
      tags:
        - Issues
      summary: Assign issue owner
      operationId: assignIssueOwner
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignOwnerRequest'
      responses:
        '200':
          description: Owner assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '404':
          description: Issue not found
  /issues/{issueId}/actions/close:
    post:
      tags:
        - Issues
      summary: Close an issue
      operationId: closeIssue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Issue closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '409':
          description: Issue not ready to close
  /issues/{issueId}/actions/abandon:
    post:
      tags:
        - Issues
      summary: Abandon an issue
      operationId: abandonIssue
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Issue abandoned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '404':
          description: Issue not found
  /issues/{issueId}/phases:
    post:
      tags:
        - Issues
      summary: Add a phase to an issue
      operationId: addPhase
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPhaseRequest'
      responses:
        '200':
          description: Phase added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '404':
          description: Issue not found
  /issues/{issueId}/phases/{phaseId}:
    patch:
      tags:
        - Issues
      summary: Update phase
      operationId: updatePhase
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
        - name: phaseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePhaseRequest'
      responses:
        '200':
          description: Phase updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '404':
          description: Issue or phase not found
  /issues/{issueId}/phases/{phaseId}/assignee:
    patch:
      tags:
        - Issues
      summary: Assign a phase assignee
      operationId: assignPhaseAssignee
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
        - name: phaseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignAssigneeRequest'
      responses:
        '200':
          description: Assignee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '404':
          description: Issue or phase not found
  /issues/{issueId}/phases/{phaseId}/actions/fail:
    post:
      tags:
        - Issues
      summary: Mark phase as failed
      operationId: failPhase
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
        - name: phaseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Phase failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '409':
          description: Invalid transition
  /issues/{issueId}/phases/{phaseId}/actions/reopen:
    post:
      tags:
        - Issues
      summary: Reopen phase
      operationId: reopenPhase
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
        - name: phaseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Phase reopened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '409':
          description: Invalid transition
  /issues/{issueId}/phases/{phaseId}/tasks:
    post:
      tags:
        - Issues
      summary: Add a task to a phase
      operationId: addTask
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
        - name: phaseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTaskRequest'
      responses:
        '200':
          description: Task added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '404':
          description: Issue or phase not found
  /issues/{issueId}/phases/{phaseId}/tasks/{taskId}:
    patch:
      tags:
        - Issues
      summary: Update task
      operationId: updateTask
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
        - name: phaseId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetail'
        '404':
          description: Issue, phase, or task not found
  /issues/{issueId}/history:
    get:
      tags:
        - Issues
      summary: Get issue history
      operationId: getIssueHistory
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Issue history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueHistoryResponse'
        '404':
          description: Issue not found
  /issues/{issueId}/comments:
    get:
      tags:
        - Issues
      summary: List issue comments
      operationId: listIssueComments
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comment thread
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueCommentsResponse'
    post:
      tags:
        - Issues
      summary: Add issue comment
      operationId: addIssueComment
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCommentRequest'
      responses:
        '201':
          description: Comment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
  /issues/{issueId}/comments/read:
    post:
      tags:
        - Issues
      summary: Mark issue comments as read
      operationId: markIssueCommentsRead
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentReadRequest'
      responses:
        '200':
          description: Read state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentReadResponse'
...
components:
  schemas:
    CreateIssueRequest:
      type: object
      required:
        - title
        - ownerId
        - projectId
        - description
        - phases
      properties:
        title:
          type: string
        ownerId:
          type: string
        projectId:
          type: string
        description:
          type: string
        deadline:
          type: string
          format: date
        phases:
          type: array
          items:
            $ref: '#/components/schemas/CreatePhaseRequest'
    UpdateIssueRequest:
      type: object
      properties:
        title:
          type: string
        ownerId:
          type: string
        projectId:
          type: string
        description:
          type: string
        deadline:
          type: string
          format: date
    AssignOwnerRequest:
      type: object
      required:
        - ownerId
      properties:
        ownerId:
          type: string
    AssignAssigneeRequest:
      type: object
      required:
        - assigneeId
      properties:
        assigneeId:
          type: string
    AddPhaseRequest:
      type: object
      required:
        - name
        - assigneeId
      properties:
        name:
          type: string
        assigneeId:
          type: string
        kind:
          $ref: '#/components/schemas/PhaseKind'
        deadline:
          type: string
          format: date
    UpdatePhaseRequest:
      type: object
      properties:
        name:
          type: string
        assigneeId:
          type: string
        status:
          $ref: '#/components/schemas/PhaseStatus'
        completionComment:
          type: string
        completionArtifactUrl:
          type: string
          format: uri
        kind:
          $ref: '#/components/schemas/PhaseKind'
        deadline:
          type: string
          format: date
    AddTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        assigneeId:
          type: string
        startDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/TaskDependency'
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        assigneeId:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        startDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/TaskDependency'
...
    IssueListItem:
      type: object
      required:
        - id
        - title
        - ownerId
        - phaseCount
        - status
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        ownerId:
          type: string
        projectId:
          type: string
        phaseCount:
          type: integer
          format: int32
        phaseStatusCounts:
          type: object
          additionalProperties:
            type: integer
            format: int32
        phaseProgress:
          type: array
          items:
            $ref: '#/components/schemas/PhaseProgressItem'
        status:
          $ref: '#/components/schemas/IssueStatus'
        deadline:
          type: string
          format: date
    IssueDetail:
      type: object
      required:
        - id
        - title
        - ownerId
        - description
        - phases
        - status
        - version
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        ownerId:
          type: string
        projectId:
          type: string
        phases:
          type: array
          items:
            $ref: '#/components/schemas/PhaseResponse'
        status:
          $ref: '#/components/schemas/IssueStatus'
        deadline:
          type: string
          format: date
        version:
          type: integer
          format: int64
        allowedActions:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ActionDecision'
...
    IssueFacetResponse:
      type: object
      required:
        - owners
        - assignees
        - projects
        - statuses
        - phaseKinds
      properties:
        owners:
          type: array
          items:
            $ref: '#/components/schemas/IssueFacetOption'
        assignees:
          type: array
          items:
            $ref: '#/components/schemas/IssueFacetOption'
        projects:
          type: array
          items:
            $ref: '#/components/schemas/IssueFacetOption'
        statuses:
          type: array
          items:
            $ref: '#/components/schemas/IssueFacetOption'
        phaseKinds:
          type: array
          items:
            $ref: '#/components/schemas/IssueFacetOption'
...
    IssueHistoryResponse:
      type: object
      required:
        - activity
        - audit
      properties:
        activity:
          type: array
          items:
            $ref: '#/components/schemas/ActivityEntry'
        audit:
          type: array
          items:
            $ref: '#/components/schemas/AuditEntry'
...
    AddCommentRequest:
      type: object
      required:
        - body
      properties:
        body:
          type: string
        mentions:
          type: array
          items:
            type: string
...
    IssueCommentsResponse:
      type: object
      required:
        - comments
        - unreadCount
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'
        unreadCount:
          type: integer
        lastReadAt:
          type: string
          format: date-time
        latestCommentAt:
          type: string
          format: date-time
        firstUnreadCommentId:
          type: string
    CommentReadRequest:
      type: object
      properties:
        lastReadCommentId:
          type: string
    CommentReadResponse:
      type: object
      required:
        - lastReadAt
        - unreadCount
      properties:
        lastReadAt:
          type: string
          format: date-time
        unreadCount:
          type: integer
        latestCommentAt:
          type: string
          format: date-time
...
    IssueStatus:
      type: string
      enum:
        - CREATED
        - NOT_ACTIVE
        - IN_ANALYSIS
        - IN_DEVELOPMENT
        - IN_TEST
        - IN_ROLLOUT
        - DONE
        - FAILED
    PhaseStatus:
      type: string
      enum:
        - NOT_STARTED
        - IN_PROGRESS
        - DONE
        - FAILED
...
    PhaseKind:
      type: string
      enum:
        - INVESTIGATION
        - PROPOSE_SOLUTION
        - DEVELOPMENT
        - ACCEPTANCE_TEST
        - ROLLOUT